requirements.txt
bs4
python-dotenv
aiohttp==3.8.1
